Almost finished Http Delete

+++ Fittify.DbResetter/DbConnection.cs
@@ -1,7 +1,10 @@
 ﻿using System;
+using System.Data;
 using System.Data.SqlClient;
 using System.IO;
+using System.Threading;
 using Fittify.Test.Core.Seed;
+using Microsoft.EntityFrameworkCore.Storage;
 using Microsoft.Extensions.Configuration;
 
 namespace Fittify.DbResetter
@@ -41,24 +44,51 @@ namespace Fittify.DbResetter
             using (var master = new SqlConnection(GetMasterConnectionStringFromAppsettingsJson()))
             {
                 string fittifyDbName;
-                using (var fittify = new SqlConnection(GetFittifyConnectionStringFromAppsettingsJson()))
+                var connectionString = GetFittifyConnectionStringFromAppsettingsJson();
+                using (var fittify = new SqlConnection(connectionString))
                 {
                     fittifyDbName = fittify.Database;
                 }
-                master.Open();
 
-                using (var command = master.CreateCommand())
+                //master.Open();
+
+                try
                 {
-                    // SET SINGLE_USER will close any open connections that would prevent the drop
-                    command.CommandText
-                        = string.Format(@"IF EXISTS (SELECT * FROM sys.databases WHERE name = N'{0}')
-                                        BEGIN
-                                            ALTER DATABASE [{0}] SET SINGLE_USER WITH ROLLBACK IMMEDIATE;
-                                            DROP DATABASE [{0}];
-                                        END", fittifyDbName);
-
-                    command.ExecuteNonQuery();
+                    master.Open();
                 }
+                catch (Exception e)
+                {
+                    Console.WriteLine(fittifyDbName + "doesn't seem to exist");
+                }
+
+                if (master.State == ConnectionState.Open)
+                {
+                    using (var command = master.CreateCommand())
+                    {
+                        //command.CommandText =
+                        //    string.Format(@"USE master;
+                        //                    --GO
+                        //                    IF EXISTS (SELECT * FROM sys.databases WHERE name = N'{0}')
+                        //                    BEGIN
+                        //                    ALTER DATABASE {0} SET SINGLE_USER WITH ROLLBACK IMMEDIATE;
+                        //                    DROP DATABASE {0};
+                        //                    END", fittifyDbName);
+
+                        //command.ExecuteNonQuery();
+
+                        //USED TO WORK
+                        // SET SINGLE_USER will close any open connections that would prevent the drop
+                        command.CommandText
+                            = string.Format(@"IF EXISTS (SELECT * FROM sys.databases WHERE name = N'{0}')
+                                            BEGIN
+                                                ALTER DATABASE [{0}] SET SINGLE_USER WITH ROLLBACK IMMEDIATE;
+                                                DROP DATABASE [{0}];
+                                            END", fittifyDbName);
+
+                        command.ExecuteNonQuery();
+                    }
+                }
+                //Thread.Sleep(20000);
             }
         }
